+++ LughSharp/Lugh/Utils/ScreenUtils.cs
@@ -23,7 +23,6 @@
 // ///////////////////////////////////////////////////////////////////////////////
 
 using LughSharp.Lugh.Graphics;
-using LughSharp.Lugh.Graphics.G2D;
 using LughSharp.Lugh.Graphics.Images;
 using LughSharp.Lugh.Graphics.OpenGL;
 using LughSharp.Lugh.Maths;
@@ -32,25 +31,25 @@ using LughSharp.Lugh.Utils.Buffers;
 namespace LughSharp.Lugh.Utils;
 
 /// <summary>
-/// Class with static helper methods related to currently bound OpenGL frame buffer,
-/// including access to the current OpenGL FrameBuffer. These methods can be used to
-/// get the entire screen content or a portion thereof.
+///     Class with static helper methods related to currently bound OpenGL frame buffer,
+///     including access to the current OpenGL FrameBuffer. These methods can be used to
+///     get the entire screen content or a portion thereof.
 /// </summary>
 [PublicAPI]
 public class ScreenUtils
 {
     /// <summary>
-    /// Returns the current framebuffer contents as a <see cref="TextureRegion"/> with
-    /// a width and height equal to the current screen size.
-    /// <para>
-    /// The base <see cref="Texture"/> always has <see cref="MathUtils.NextPowerOfTwo"/>
-    /// dimensions and RGBA8888 <see cref="PixelType.Format"/>. It can be accessed via
-    /// <see cref="TextureRegion.Texture"/>. The texture is not managed and has to be
-    /// reloaded manually on a context loss.
-    /// </para>
-    /// <para>
-    /// The returned TextureRegion is flipped along the Y axis by default.
-    /// </para>
+    ///     Returns the current framebuffer contents as a <see cref="TextureRegion" /> with
+    ///     a width and height equal to the current screen size.
+    ///     <para>
+    ///         The base <see cref="Texture" /> always has <see cref="MathUtils.NextPowerOfTwo" />
+    ///         dimensions and RGBA8888 <see cref="PixelType.Format" />. It can be accessed via
+    ///         <see cref="TextureRegion.Texture" />. The texture is not managed and has to be
+    ///         reloaded manually on a context loss.
+    ///     </para>
+    ///     <para>
+    ///         The returned TextureRegion is flipped along the Y axis by default.
+    ///     </para>
     /// </summary>
     public static TextureRegion FrameBufferTexture
     {
@@ -69,7 +68,7 @@ public class ScreenUtils
     }
 
     /// <summary>
-    /// Clears the color buffers and optionally the depth buffer.
+    ///     Clears the color buffers and optionally the depth buffer.
     /// </summary>
     /// <param name="color">Color to clear the color buffers with.</param>
     /// <param name="clearDepth">Clears the depth buffer if true.</param>
@@ -79,7 +78,7 @@ public class ScreenUtils
     }
 
     /// <summary>
-    /// Clears the color buffers and optionally the depth buffer.
+    ///     Clears the color buffers and optionally the depth buffer.
     /// </summary>
     /// <param name="a"> Alpha component. </param>
     /// <param name="clearDepth"> Clears the depth buffer if true.</param>
@@ -90,7 +89,7 @@ public class ScreenUtils
     {
         GdxApi.Bindings.ClearColor( r, g, b, a );
 
-        var mask = ( uint ) IGL.GL_COLOR_BUFFER_BIT;
+        var mask = ( uint )IGL.GL_COLOR_BUFFER_BIT;
 
         if ( clearDepth )
         {
@@ -101,21 +100,21 @@ public class ScreenUtils
     }
 
     /// <summary>
-    /// Returns a portion of the current framebuffer contents specified by x, y,
-    /// width and height as a <see cref="TextureRegion"/> with the same dimensions.
-    /// <para>
-    /// The base <see cref="Texture"/> always has <see cref="MathUtils.NextPowerOfTwo"/>
-    /// dimensions and RGBA8888 <see cref="PixelType.Format"/>. It can be accessed via
-    /// <see cref="TextureRegion.Texture"/>.
-    /// </para>
-    /// <para>
-    /// This texture is not managed and has to be reloaded manually on a context loss.
-    /// If the width and height specified are larger than the framebuffer dimensions,
-    /// the Texture will be padded accordingly.
-    /// </para>
-    /// <para>
-    /// Pixels that fall outside of the current screen will have RGBA values of 0.
-    /// </para>
+    ///     Returns a portion of the current framebuffer contents specified by x, y,
+    ///     width and height as a <see cref="TextureRegion" /> with the same dimensions.
+    ///     <para>
+    ///         The base <see cref="Texture" /> always has <see cref="MathUtils.NextPowerOfTwo" />
+    ///         dimensions and RGBA8888 <see cref="PixelType.Format" />. It can be accessed via
+    ///         <see cref="TextureRegion.Texture" />.
+    ///     </para>
+    ///     <para>
+    ///         This texture is not managed and has to be reloaded manually on a context loss.
+    ///         If the width and height specified are larger than the framebuffer dimensions,
+    ///         the Texture will be padded accordingly.
+    ///     </para>
+    ///     <para>
+    ///         Pixels that fall outside of the current screen will have RGBA values of 0.
+    ///     </para>
     /// </summary>
     /// <param name="x"> the x position of the framebuffer contents to capture </param>
     /// <param name="y"> the y position of the framebuffer contents to capture </param>
@@ -140,18 +139,18 @@ public class ScreenUtils
         return textureRegion;
     }
 
-    /// <inheritdoc cref="Pixmap.CreateFromFrameBuffer"/> 
+    /// <inheritdoc cref="Pixmap.CreateFromFrameBuffer" />
     public static Pixmap GetFrameBufferPixmap( int x, int y, int w, int h )
     {
         return Pixmap.CreateFromFrameBuffer( x, y, w, h );
     }
 
     /// <summary>
-    /// Returns the current framebuffer contents as a byte[] array with a length equal
-    /// to screen width * height * 4. The byte[] will always contain RGBA8888 data.
-    /// Because of differences in screen and image origins the framebuffer contents should
-    /// be flipped along the Y axis if you intend save them to disk as a bitmap. Flipping
-    /// is not a cheap operation, so use this functionality wisely.
+    ///     Returns the current framebuffer contents as a byte[] array with a length equal
+    ///     to screen width * height * 4. The byte[] will always contain RGBA8888 data.
+    ///     Because of differences in screen and image origins the framebuffer contents should
+    ///     be flipped along the Y axis if you intend save them to disk as a bitmap. Flipping
+    ///     is not a cheap operation, so use this functionality wisely.
     /// </summary>
     /// <param name="flipY"> whether to flip pixels along Y axis</param>
     public static byte[] GetFrameBufferPixels( bool flipY )
@@ -168,16 +167,16 @@ public class ScreenUtils
     }
 
     /// <summary>
-    /// Returns a portion of the current framebuffer contents specified by x, y, width and height,
-    /// as a byte[] array with a length equal to the specified width * height * 4. The byte[] will
-    /// always contain RGBA8888 data. If the width and height specified are larger than the framebuffer
-    /// dimensions, the Texture will be padded accordingly. Pixels that fall outside of the current
-    /// screen will have RGBA values of 0. Because of differences in screen and image origins the
-    /// framebuffer contents should be flipped along the Y axis if you intend save them to disk as
-    /// a bitmap.
-    /// <para>
-    /// Flipping is not a cheap operation, so use this functionality wisely.
-    /// </para>
+    ///     Returns a portion of the current framebuffer contents specified by x, y, width and height,
+    ///     as a byte[] array with a length equal to the specified width * height * 4. The byte[] will
+    ///     always contain RGBA8888 data. If the width and height specified are larger than the framebuffer
+    ///     dimensions, the Texture will be padded accordingly. Pixels that fall outside of the current
+    ///     screen will have RGBA values of 0. Because of differences in screen and image origins the
+    ///     framebuffer contents should be flipped along the Y axis if you intend save them to disk as
+    ///     a bitmap.
+    ///     <para>
+    ///         Flipping is not a cheap operation, so use this functionality wisely.
+    ///     </para>
     /// </summary>
     /// <param name="x"> Portion X position. </param>
     /// <param name="y"> Portion Y position. </param>
@@ -191,10 +190,10 @@ public class ScreenUtils
         GdxApi.Bindings.PixelStorei( IGL.GL_PACK_ALIGNMENT, 1 );
 
         var pixels = new ByteBuffer( numBytes );
-        
+
         fixed ( void* ptr = &pixels.ToArray()[ 0 ] )
         {
-            GdxApi.Bindings.ReadPixels( x, y, w, h, IGL.GL_RGBA, IGL.GL_UNSIGNED_BYTE, ptr );
+            GdxApi.Bindings.ReadPixels( x, y, w, h, IGL.GL_RGBA, IGL.GL_UNSIGNED_BYTE, ( IntPtr )ptr );
         }
 
         var lines = new byte[ numBytes ];
